requirements.txt
qrcode
Pillow
reportlab
pandas
